+++ src/FsPickler.Json/JsonWriter.fs
@@ -8,6 +8,8 @@ open Newtonsoft.Json
 
 open MBrace.FsPickler
 
+#nowarn "44" // BsonWriter
+
 /// <summary>
 ///     Json format serializer.
 /// </summary>